renames repositories are storage agents within persistence and updates Storage agents for all CRUD
