+++ examples/bri-3/src/bri/workgroup/worksteps/persistence/mockWorkstepsStorage.agent.ts
@@ -3,7 +3,7 @@ import { uuid } from 'uuidv4';
 import { Workstep } from '../models/workstep';
 
 @Injectable()
-export class MockWorkstepRepository {
+export class MockWorkstepStorageAgent {
   private workstepsStore: Workstep[] = [];
 
   async getWorkstepById(id: string): Promise<Workstep> {
